+++ .eslintrc.js
@@ -2,6 +2,9 @@ module.exports = {
 	extends: [
 		'@nextcloud',
 	],
+	plugins: [
+		"eslint-plugin-risxss"
+	],
 	"rules": {
         "node/no-extraneous-import": ["error", {
             "allowModules": [