+++ src/views/Timeline.vue
@@ -25,6 +25,7 @@
 		<Sidebar v-if="sidebar"
 			ref="sidebar"
 			:loading="loading"
+			:views-in-files="true"
 			@update="getData"
 			@sign:document="signDocument"
 			@closeSidebar="closeSidebar" />