+++ src/Components/Description/Description.vue
@@ -1,3 +1,4 @@
+<!-- eslint-disable vue/no-v-html -->
 <!--
 - @copyright Copyright (c) 2021 Lyseon Tech <contato@lt.coop.br>
 -
@@ -23,10 +24,10 @@
 
 <template>
 	<div class="container-desc">
-		<header>
+		<header v-show="!viewHeader">
 			<img :src="image">
 			<p>{{ t('libresign', pdfName) }}</p>
-			<span>{{ t('libresign', pdfDescription) }}</span>
+			<span v-html="markedDescription" />
 		</header>
 		<div id="body">
 			<form @submit="(e) => e.preventDefault()">
@@ -49,11 +50,17 @@
 						</a>
 						<button
 							type="button"
-							:value="buttonValue"
+							:value=" t('libresign', 'Sign the document.')"
 							:class="!updating ? 'primary' : 'primary loading'"
 							:disabled="disableButton"
 							@click="sign">
 							{{ t('libresign', 'Sign the document.') }}
+						</button>
+						<button v-show="showDoc"
+							type="button"
+							class="button secondary"
+							@click="emitShow">
+							{{ t('libresign', 'Show Document') }}
 						</button>
 					</div>
 				</div>
@@ -78,7 +85,8 @@ import CreatePassword from '../../views/CreatePassword.vue'
 import axios from '@nextcloud/axios'
 import Image from '../../assets/images/application-pdf.png'
 import { generateUrl } from '@nextcloud/router'
-import { translate as t } from '@nextcloud/l10n'
+import marked from 'marked'
+import dompurify from 'dompurify'
 
 export default {
 	name: 'Description',
@@ -113,9 +121,11 @@ export default {
 			signaturePath: '2',
 			password: '',
 			asign: true,
-			buttonValue: t('libresign', 'Sign the document.'),
 			modal: false,
 			havePfx: false,
+			showDoc: false,
+			viewHeader: false,
+			width: window.innerWidth,
 		}
 	},
 
@@ -123,8 +133,28 @@ export default {
 		hasSavePossible() {
 			return !!this.password
 		},
+		markedDescription() {
+			return dompurify.sanitize(marked(this.pdfDescription), { USE_PROFILES: { html: false } })
+		},
+	},
+	watch: {
+		width(newVal, oldVal) {
+			if (newVal <= 650) {
+				this.showDoc = true
+			}
+			if (newVal > 650) {
+				this.showDoc = false
+			}
+		},
 	},
 	created() {
+		this.$nextTick(() => {
+			window.addEventListener('resize', this.onResize)
+		})
+		this.width <= 650
+			? this.showDoc = true
+			: this.showDoc = false
+
 		this.getMe()
 	},
 
@@ -164,6 +194,15 @@ export default {
 		handleModal(status) {
 			this.modal = status
 		},
+		emitShow() {
+			this.$emit('onDocument', true)
+			this.showDoc = false
+			this.viewHeader = true
+		},
+		onResize() {
+			this.width = window.innerWidth
+		},
+
 	},
 }
 </script>