+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Types of changes:
 
 # Changed
 - PDF preview on mobile when signing
+- Only show request signing button if file not signed
 
 ## 2.4.3 - 2021-07-14