+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Types of changes:
 - Test for validation of Swagger documentation
 - GitHub action to add a changelog reminder
 - View document on mobile before sign
+- Markdown formatting for description.
 
 # Changed
 - Bump max Nextcloud version to 23